server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
pyyaml>=6.0
pydantic-settings>=2.1.0
Pillow>=10.0.0
openai>=1.10.0
weasyprint>=61.0
aiofiles>=23.2.0
